requirements.txt
Django
pytube3
